chore(endpoints): Create middleware for cleaner code[finishes #161678456]
